new city map asset models
