+++ Team10_JournalApp/Components/CityJournalMapView.swift
@@ -10,63 +10,6 @@ import SwiftUI
 //! Magic Number
 let isIphone16ProMaxPortrait: Bool = UIScreen.main.bounds.height == 956.0
 
-enum Map: String {
-    case Map1 = "Map_1"
-    case Map2 = "Map_2"
-    case Map3 = "Map_3"
-    case Map4 = "Map_4"
-    case LoadingMap = "LoadingMap"
-    case NotFoundMap = "NotFoundMap"
-}
-
-enum BuildingCategory {
-    case Building
-    case Construction
-    case Ruin
-}
-
-enum Building: String {
-    case BlueTower = "blue_building"
-    case BrownTower = "brown_building"
-    case GreenTower = "green_building"
-    case LightBlueTower = "light_blue_building"
-    case LightBrownTower = "light_brown_building"
-    case LightGreenTower = "light_green_building"
-    case RedTower = "red_building"
-    
-    case BlueConstruction = "BlueConstruction"
-    case BrownConstruction = "BrownConstruction"
-    case DarkRedConstruction = "DarkRedConstruction"
-    case GreenConstruction = "GreenConstruction"
-    case PurpleConstruction = "PurpleConstruction"
-    case RedConstruction = "RedConstruction"
-    case YellowConstruction = "YellowConstruction"
-    
-    case BlueRuin = "BlueRuin"
-    case BrownRuin = "BrownRuin"
-    case DarkRedRuin = "DarkRedRuin"
-    case GreenRuin = "GreenRuin"
-    case PurpleRuin = "PurpleRuin"
-    case RedRuin = "RedRuin"
-    case YellowRuin = "YellowRuin"
-    
-    var category: BuildingCategory {
-        switch self {
-            case .BlueTower, .BrownTower, .GreenTower, .LightBlueTower,
-                    .LightBrownTower, .LightGreenTower, .RedTower:
-                return .Building
-            
-            case .BlueConstruction, .BrownConstruction, .DarkRedConstruction, .GreenConstruction,
-                    .PurpleConstruction, .RedConstruction, .YellowConstruction:
-                return .Construction
-            
-            case .BlueRuin, .BrownRuin, .DarkRedRuin, .GreenRuin,
-                    .PurpleRuin, .RedRuin, .YellowRuin:
-                return .Ruin
-        }
-    }
-}
-
 struct BuildingConfig {
     var style: Building
     var onClick: () -> Void
